ARUHA-282: add toogle to switch off creation and deletion of event type

+++ src/main/java/de/zalando/aruha/nakadi/controller/EventTypeController.java
@@ -16,6 +16,7 @@ import de.zalando.aruha.nakadi.partitioning.PartitionResolver;
 import de.zalando.aruha.nakadi.problem.ValidationProblem;
 import de.zalando.aruha.nakadi.repository.EventTypeRepository;
 import de.zalando.aruha.nakadi.repository.TopicRepository;
+import de.zalando.aruha.nakadi.util.FeatureToggleService;
 import de.zalando.aruha.nakadi.util.UUIDGenerator;
 import org.everit.json.schema.SchemaException;
 import org.everit.json.schema.loader.SchemaLoader;
@@ -38,6 +39,8 @@ import javax.validation.Valid;
 import java.util.List;
 import java.util.Objects;
 
+import static de.zalando.aruha.nakadi.util.FeatureToggleService.Feature.DISABLE_EVENT_TYPE_CREATION;
+import static de.zalando.aruha.nakadi.util.FeatureToggleService.Feature.DISABLE_EVENT_TYPE_DELETION;
 import static org.springframework.http.ResponseEntity.status;
 import static org.zalando.problem.spring.web.advice.Responses.create;
 
@@ -51,6 +54,7 @@ public class EventTypeController {
     private final TopicRepository topicRepository;
     private final PartitionResolver partitionResolver;
     private final Enrichment enrichment;
+    private final FeatureToggleService featureToggleService;
     private final UUIDGenerator uuidGenerator;
 
     @Autowired
@@ -58,11 +62,14 @@ public class EventTypeController {
                                final TopicRepository topicRepository,
                                final PartitionResolver partitionResolver,
                                final Enrichment enrichment,
-                               final UUIDGenerator uuidGenerator) {
+                               final FeatureToggleService featureToggleService,
+                               final UUIDGenerator uuidGenerator)
+    {
         this.eventTypeRepository = eventTypeRepository;
         this.topicRepository = topicRepository;
         this.partitionResolver = partitionResolver;
         this.enrichment = enrichment;
+        this.featureToggleService = featureToggleService;
         this.uuidGenerator = uuidGenerator;
     }
 
@@ -76,7 +83,11 @@ public class EventTypeController {
     @RequestMapping(method = RequestMethod.POST)
     public ResponseEntity<?> createEventType(@Valid @RequestBody final EventType eventType,
                                              final Errors errors,
-                                             final NativeWebRequest nativeWebRequest) {
+                                             final NativeWebRequest nativeWebRequest)
+    {
+        if (featureToggleService.isFeatureEnabled(DISABLE_EVENT_TYPE_CREATION)) {
+            return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+        }
         if (errors.hasErrors()) {
             return create(new ValidationProblem(errors), nativeWebRequest);
         }
@@ -110,7 +121,11 @@ public class EventTypeController {
 
     @RequestMapping(value = "/{name:.+}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteEventType(@PathVariable("name") final String eventTypeName,
-                                             final NativeWebRequest nativeWebRequest) {
+                                             final NativeWebRequest nativeWebRequest)
+    {
+        if (featureToggleService.isFeatureEnabled(DISABLE_EVENT_TYPE_DELETION)) {
+            return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+        }
         try {
             EventType eventType = eventTypeRepository.findByName(eventTypeName);
             eventTypeRepository.removeEventType(eventTypeName);
